requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
python-dotenv
groq
pydantic
